Refactor downstream options
